mongo: perform a hostname fixup on repset primaries
When rs.initiate() is run without previously defining a replication set
config, the primary node does a lookup of its own hostname. When
running in EC2 this lookup returns the ip-xxx-xxx-xxx-xxx.ec2.internal
hostname instead of the public ec2-xxx...compute.amazonaws.com hostname

This results various failures, the most obvious of which is sharding
initialization failure. Maybe there is a better fix, but reassigning
the hostname to match the format used by the rest of the nodes resolves
the issue just as well.