mongo: preserve selinux contexts
This playbook assumes it is running on an RHEL6 (or equivalent ie.
CentOS 6) system. Such boxes come with SELinux enabled by default.
Without the python selinux bindings installed, copying files and/or
installing templates results in the clobbering of existing SELinux
contexts on those files. This resulted in the inability to reconfigure
iptables among other issues. Since ansible is awesome, it does the
right thing (preserves security contexts) once the python bindings are
installed.