mongo: only populate hostfile with mongo hosts
the "mongo_servers" group contains all mongo servers currently, so we
can use it instead of the "all" group. There is no guarantee that hosts
in the "all" group will have an eth0 interface which causes this step
to fail. I may remove this step entirely since our hostnames are
resolvable